fix a bug of runSinglePlugin

+++ eBPF_Visualization/eBPF_server/service/ebpfplugins/ebpfplugins.go
@@ -87,11 +87,7 @@ func (ebpf *EbpfpluginsService) LoadEbpfPlugins(e request.PluginInfo) (err error
 	}
 
 	// 2.加载执行
-	err = runSinglePlugin(plugin.PluginPath)
-	if err != nil {
-		return err
-	}
-
+	go runSinglePlugin(plugin.PluginPath)
 	// 3.执行之后结果执行，成功还是失败
 	plugin.State = 1 // 表示已经成功加载内核中运行
 	err = global.GVA_DB.Save(plugin).Error
@@ -115,9 +111,14 @@ func (ebpf *EbpfpluginsService) UnloadEbpfPlugins(e request.PluginInfo) (err err
 	var plugin ebpfplugins.EbpfPlugins
 	db.Where("id = ?", e.PluginId).First(&plugin)
 	if plugin.State == 1 {
-		plugin.State = 0
-		killProcess(plugin.PluginPath)
-		err = global.GVA_DB.Save(plugin).Error
+		if pluginPid[plugin.PluginPath] == 0 {
+			plugin.State = 0
+			err = global.GVA_DB.Save(plugin).Error
+		} else {
+			plugin.State = 0
+			killProcess(plugin.PluginPath)
+			err = global.GVA_DB.Save(plugin).Error
+		}
 	}
 
 	return err
